+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.NEXT
+## mm/dd/2022
+
+1. [](#new)
+   * Added support for `multipart/form-data` content type in PUT and PATCH requests
+   * Added support for flex object relationships
+
 # v1.7.34
 ## 06/14/2022